Introduce feeCollectWeight logic
After calling reinvest(), the realX and realY values of a position can slightly decrease because of the reinvest bounty paid to the operator.

This could cause a healthy position to become liquidatable right after reinvest() is called. Therefore, liquidators have an incentive to call reinvest() before liquidating a position in order to be more competitive. This reduces the efficiency of liquidations and will make liquidator contracts more complex.

To fix the issue, for liquidations we pessimistically use a value lower than the real one for feeCollected. We lower it by 5%, which is more than the expected cost of reinvest().

Notice: the value is lowered just for lowestPrice and highestPrice, while it remains the same for currentPrice. Only lowest and highest are used to check if a position isLiquidatable. On the other hand current price is used to calculate isUnderwater and the seizePercentage, so it shouldn't be lowered.