Introduce blockOfLastRestructureOrLiquidation
This fixes 2 issues:

1. In some edge scenarios, a position is not liquidatable right after restructure bad debt is called.
You can reproduce this by creating a one sided UniV3 collateral position, and borrow the same side that you're using as collateral. In this specific scenario, it turns out that: isLiquidatable == isUnderwater. Therefore liquidate() would never work because a position can't be liquidatable without being underwater. A malicious user could use this bug to lock the capital of the lenders.
This update introduces a fix. A position can always be liquidated if restructureBadDebt is called in the same block.

2. Often, it is impossible to liquidate a position 100% because after liquidating one side the position enters in a healty state. This is suboptimal because it creates more complexity for the liquidators.
Similarly to 1. this update establishes that a position can alwyas be liquidated if it has already been liquidated in the same block.