Enhance CollateralMath precision
Increased the precision of "getValue" by 32 bit.

Given the nature of the lending math, it would technically be possible to borrow dust amounts while having 0 collateral deposited. That's possible because getDebtValue can return 0 if debtX and debtY are low amounts.

Increasing the precision largely decrease the likelihood of that happening.